Fix integer cast error

+++ src/main/java/com/example/demo/services/CandlestickCheckService.java
@@ -16,6 +16,6 @@ public class CandlestickCheckService {
     }
 
     private boolean checkIfCandleGreen(Candlestick candle) {
-        return Integer.parseInt(candle.getHigh()) > Integer.parseInt(candle.getOpen());
+        return Double.parseDouble(candle.getHigh()) > Double.parseDouble(candle.getOpen());
     }
 }
